requirements.txt
bs4
ipykernel
python-dotenv
faiss-cpu
langchain
langchain-groq
langchain-core
langchain-openai
langchain-community
pypdf
arxiv
pymupdf
wikipedia
transformers
langchain_huggingface
sentence-transformers
langgraph
langgraph-cli[inmem]